+++ index.ts
@@ -180,15 +180,15 @@ export const consumeAllItemsAndroid = (): Promise<string[]> =>
 /**
  * Get a list of products (consumable and non-consumable items, but not subscriptions)
  * @param {string[]} skus The item skus
- * @returns {Promise<InAppPurchase[]>}
+ * @returns {Promise<Product[]>}
  */
-export const getProducts = (skus: string[]): Promise<InAppPurchase[]> =>
+export const getProducts = (skus: string[]): Promise<Product[]> =>
   Platform.select({
     ios: async () => {
       if (!RNIapIos) {
         return [];
       }
-      return RNIapIos.getItems(skus).then((items) =>
+      return RNIapIos.getItems(skus).then((items: Product[]) =>
         items.filter((item) => item.productId),
       );
     },
@@ -203,15 +203,13 @@ export const getProducts = (skus: string[]): Promise<InAppPurchase[]> =>
 /**
  * Get a list of subscriptions
  * @param {string[]} skus The item skus
- * @returns {Promise<SubscriptionPurchase[]>}
+ * @returns {Promise<Subscription[]>}
  */
-export const getSubscriptions = (
-  skus: string[],
-): Promise<SubscriptionPurchase[]> =>
+export const getSubscriptions = (skus: string[]): Promise<Subscription[]> =>
   Platform.select({
     ios: async () => {
       checkNativeiOSAvailable();
-      return RNIapIos.getItems(skus).then((items) =>
+      return RNIapIos.getItems(skus).then((items: Subscription[]) =>
         items.filter((item) => skus.includes(item.productId)),
       );
     },