+++ CHANGELOG.md
@@ -27,7 +27,8 @@
   - rc7
     - Fix optional parameters not specified.
   - rc8
-    - Pu promises returns back in `android` for users who still need it.
+    - Put promises returns back in `android` for users who still need it.
+    - Fixed typings for `getProducts` and `getSubscriptions`.
 - **[3.5.+]**
   - Migrated to `ts` project. Now we only need to concentrate on `index.ts` file. `index.js`, `index.js.flow`, `index.d.ts` is generated during build.
   - `consumeAllItems` returns consumed tokens instead of void.